chore: Re-add `ErrorBoundary`
- https://react.dev/reference/react/Component#catching-rendering-errors-with-an-error-boundary
